+++ packages/openjpeg/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.5](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-openjpeg@1.3.4...@cornerstonejs/codec-openjpeg@1.3.5) (2026-09-02)
+
+
+### Bug Fixes
+
+* **openjpeg:** correct the buffer-stream skip signature and decoded image dimensions (#63) ([1285c02](https://github.com/cornerstonejs/codecs/commit/1285c024adebe036503117b8d17e66eb23b7d18a))
+
+
+
+
+## [1.3.4](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-openjpeg@1.3.3...@cornerstonejs/codec-openjpeg@1.3.4) (2026-09-02)
+
+
+### Bug Fixes
+
+* consolidated codec correctness fixes (supersedes #71) (#73) ([21d4749](https://github.com/cornerstonejs/codecs/commit/21d47493be02e70067aaf6b3f23444e73d4cf37e))
+
+
+
+
+## [1.3.3](https://github.com/cornerstonejs/codecs/compare/@cornerstonejs/codec-openjpeg@1.3.2...@cornerstonejs/codec-openjpeg@1.3.3) (2026-09-01)
+
+**Note:** Version bump only for package @cornerstonejs/codec-openjpeg
+
+
+
+
+## [1.3.2](https://localhost/compare/@cornerstonejs/codec-openjpeg@1.3.1...@cornerstonejs/codec-openjpeg@1.3.2) (2026-08-13)
+
+**Note:** Version bump only for package @cornerstonejs/codec-openjpeg
+
+
+
+
+
 ## [1.3.1](https://localhost/compare/@cornerstonejs/codec-openjpeg@1.3.0...@cornerstonejs/codec-openjpeg@1.3.1) (2026-07-09)